backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.2
httpx==0.27.0
python-multipart==0.0.9
slowapi==0.1.9
pytest==8.0.0
pytest-asyncio==0.23.4
